Toggle Hook added 🔨
