+++ src/ignitus-Shared/ignitus-Utilities/reactHooks/toogleHook.ts
@@ -1,0 +1,15 @@
+import React from 'react';
+
+type UseToggleReturn = [boolean, () => void];
+
+const useToggle = (inititalValue = false): UseToggleReturn => {
+  const [value, setValue] = React.useState<boolean>(inititalValue);
+
+  const toggleValue = () => {
+    setValue(!value);
+  };
+
+  return [value, toggleValue];
+};
+
+export default useToggle;
