+++ src/ignitus-UserInterfaceKit/Components/sideNavigation.tsx
@@ -1,0 +1,1 @@
+export default{}
